+++ views/login/google.ejs
@@ -30,7 +30,7 @@
   </div>
 
   <form action="/login/google" method="POST">
-    <%- include('partials/csrf') %>
+    <%- include('../partials/csrf') %>
     <input name="token">
   </form>
